app/requirements.txt
cassandra-driver
venv-pack
pyarrow
fastparquet
pathvalidate
pyspark==3.5.4
tqdm